Move endpoint APIs to their own internal package
