+++ robozonky-api/src/main/java/com/github/robozonky/internal/remote/OAuth.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 The RoboZonky Project
+ * Copyright 2020 The RoboZonky Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,10 +16,10 @@
 
 package com.github.robozonky.internal.remote;
 
-import com.github.robozonky.api.remote.ZonkyOAuthApi;
-import com.github.robozonky.api.remote.entities.ZonkyApiToken;
-
 import static com.github.robozonky.api.remote.entities.ZonkyApiToken.REFRESH_TOKEN_STRING;
+
+import com.github.robozonky.api.remote.entities.ZonkyApiToken;
+import com.github.robozonky.internal.remote.endpoints.ZonkyOAuthApi;
 
 public class OAuth {
