+++ robozonky-api/src/main/java/com/github/robozonky/internal/remote/endpoints/ReservationApi.java
@@ -14,36 +14,28 @@
  * limitations under the License.
  */
 
-package com.github.robozonky.api.remote;
+package com.github.robozonky.internal.remote.endpoints;
 
-import java.util.List;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 
-import com.github.robozonky.api.remote.entities.Investment;
-import com.github.robozonky.api.remote.entities.SellInfo;
-import com.github.robozonky.api.remote.entities.Statistics;
+import com.github.robozonky.api.remote.entities.ReservationPreferences;
+import com.github.robozonky.api.remote.entities.Reservations;
 import com.github.robozonky.internal.ApiConstants;
 import com.github.robozonky.internal.Defaults;
 
 @Produces(Defaults.MEDIA_TYPE)
 @Consumes(Defaults.MEDIA_TYPE)
-public interface PortfolioApi extends EntityCollectionApi<Investment> {
+public interface ReservationApi {
 
-    @Path(ApiConstants.INVESTMENTS)
     @GET
-    @Override
-    List<Investment> items();
+    @Path(ApiConstants.RESERVATIONS)
+    Reservations items();
 
-    @Path("/statistics/me/public-overview")
     @GET
-    Statistics getStatistics();
-
-    @Path(ApiConstants.INVESTMENTS + "/{investmentId}/smpSellInfo")
-    @GET
-    SellInfo getSellInfo(@PathParam("investmentId") long investmentId);
+    @Path(ApiConstants.RESERVATION_PREFERENCES)
+    ReservationPreferences preferences();
 
 }
