+++ robozonky-api/src/main/java/com/github/robozonky/internal/remote/endpoints/ZonkyOAuthApi.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 The RoboZonky Project
+ * Copyright 2020 The RoboZonky Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package com.github.robozonky.api.remote;
-
-import com.github.robozonky.api.remote.entities.ZonkyApiToken;
+package com.github.robozonky.internal.remote.endpoints;
 
 import javax.ws.rs.DefaultValue;
 import javax.ws.rs.FormParam;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
+
+import com.github.robozonky.api.remote.entities.ZonkyApiToken;
 
 @Path("/oauth/token")
 public interface ZonkyOAuthApi {
