+++ robozonky-api/src/test/java/com/github/robozonky/internal/remote/ProxyFactoryTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 The RoboZonky Project
+ * Copyright 2020 The RoboZonky Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,11 +16,12 @@
 
 package com.github.robozonky.internal.remote;
 
-import com.github.robozonky.api.remote.LoanApi;
+import static org.assertj.core.api.Assertions.*;
+
 import org.jboss.resteasy.client.jaxrs.ResteasyClient;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import com.github.robozonky.internal.remote.endpoints.LoanApi;
 
 class ProxyFactoryTest {
